Data.Vector.Generic.Mutable.Sized.Internal

Documentation

newtype MVector v (n :: Nat) s a

A wrapper to tag mutable vectors with a type level length.

Be careful when using the constructor here to not construct sized vectors which have a different length than that specified in the type parameter!

Constructors

MVector (v s a)