Data.Vector.Generic.Sized.Internal

Documentation

newtype Vector v (n :: Nat) a

A wrapper to tag vectors with a type level length.

Be careful when using the constructor here to not construct sized vectors which have a different length than that specified in the type parameter!

Constructors

Vector (v a)